services/college_service/requirements.txt
anyio==4.12.1
asyncpg==0.31.0
alembic==1.18.4
click==8.3.1
email-validator==2.2.0
fastapi==0.135.1
greenlet==3.3.2
grpcio==1.78.0
grpcio-tools==1.78.0
h11==0.16.0
httpx==0.28.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
protobuf==6.33.5
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dotenv==1.2.2
python-jose[cryptography]==3.5.0
python-multipart==0.0.22
SQLAlchemy==2.0.48
starlette==0.52.1
typing_extensions==4.15.0
uvicorn==0.41.0
